+++ CHANGES.rst
@@ -4,6 +4,8 @@
 - Add official support for python 3.5
 - Implement list_huisnummers_by_perceel. (#56)
 - Implement get_postadres_by_huisnummer and get_postadres_by_subadres. (#57)
+- A a property Perceel.postadressen to get the postadressen for a certain
+  Perceel. (#58)
 
 0.6.0 (01-06-2015)
 ------------------